Set IDs to 1 when no items exist
- Collections were previously undefined when no items existed, but
this was changed in commit 16ecaaf. That resulted in the ID for the
first item created of a given type to be set to 0. Our checks for the
presence of number fields on items before saving fail if it's zero, so
we were unable to associate products with a catgory with ID 0.